fix reduce as bug
