+++ paddle/cinn/hlir/dialect/operator/transforms/pd_to_cinn_pass.cc
@@ -1212,162 +1212,6 @@ class GatherOpPattern
   }
 };
 
-class ReduceAsOpPattern
-    : public pir::OpRewritePattern<paddle::dialect::ReduceAsOp> {
- public:
-  using pir::OpRewritePattern<paddle::dialect::ReduceAsOp>::OpRewritePattern;
-
-  bool MatchAndRewrite(paddle::dialect::ReduceAsOp op,
-                       pir::PatternRewriter &rewriter) const override {
-    auto x_shape =
-        phi::vectorize(op->operand_source(0)
-                           .type()
-                           .dyn_cast<paddle::dialect::DenseTensorType>()
-                           .dims());
-
-    auto y_shape =
-        phi::vectorize(op->operand_source(1)
-                           .type()
-                           .dyn_cast<paddle::dialect::DenseTensorType>()
-                           .dims());
-
-    size_t x_rank = x_shape.size();
-    size_t y_rank = y_shape.size();
-    int64_t compare_offset = x_rank - y_rank;
-    std::vector<int64_t> reduce_axis;
-
-    std::vector<int64_t> squeeze_axis;
-    for (int64_t i = 0; i < compare_offset; ++i) {
-      reduce_axis.push_back(i);
-      squeeze_axis.push_back(i);
-    }
-
-    bool x_y_shape_equal = false;
-    std::vector<symbol::DimExpr> output_dims;
-    bool is_static_shape = IsStaicShape(x_shape, y_shape);
-    if (is_static_shape) {
-      x_y_shape_equal = (x_shape == y_shape);
-      ProcessStaticShape(x_shape, y_shape, &reduce_axis);
-    } else {
-      bool can_repalce =
-          ProcessDynamicShape(op, &reduce_axis, &output_dims, &x_y_shape_equal);
-      if (!can_repalce) {
-        return true;
-      }
-    }
-    if (x_y_shape_equal) {
-      rewriter.ReplaceAllUsesWith(op.result(0), op.operand_source(0));
-      return true;
-    }
-
-    auto pir_dtype =
-        op->operand_source(0).type().dyn_cast<pir::DenseTensorType>().dtype();
-    auto phi_dtype = paddle::dialect::TransToPhiDataType(pir_dtype);
-    auto sum_op = rewriter.Build<paddle::dialect::SumOp>(
-        op.operand_source(0), reduce_axis, phi_dtype, true);
-
-    auto new_output = sum_op.result(0);
-
-    if (!is_static_shape) {
-      auto &shape_analysis =
-          pir::ShapeAnalysisManager::Instance().Get(op->GetParentProgram());
-      shape_analysis.SetShapeOrDataForValue(
-          new_output,
-          symbol::ShapeOrDataDimExprs{
-              symbol::TensorShapeOrDataDimExprs(output_dims)});
-    }
-
-    if (squeeze_axis.size() > 0) {
-      new_output =
-          rewriter.Build<paddle::dialect::SqueezeOp>(new_output, squeeze_axis)
-              .result(0);
-      if (!is_static_shape) {
-        auto &shape_analysis =
-            pir::ShapeAnalysisManager::Instance().Get(op->GetParentProgram());
-        shape_analysis.SetShapeOrDataForValue(
-            new_output,
-            shape_analysis.GetShapeOrDataForValue(op->operand_source(1)));
-      }
-    }
-
-    rewriter.ReplaceAllUsesWith(op.result(0), new_output);
-
-    rewriter.EraseOp(op);
-
-    return true;
-  }
-
- private:
-  bool IsStaicShape(const std::vector<int64_t> &x_shape,
-                    const std::vector<int64_t> &y_shape) const {
-    bool x_has_dynamic_shape =
-        std::find(x_shape.begin(), x_shape.end(), -1) != x_shape.end();
-    bool y_has_dynamic_shape =
-        std::find(y_shape.begin(), y_shape.end(), -1) != y_shape.end();
-
-    return (!x_has_dynamic_shape) && (!y_has_dynamic_shape);
-  }
-
-  void ProcessStaticShape(const std::vector<int64_t> &x_shape,
-                          const std::vector<int64_t> &y_shape,
-                          std::vector<int64_t> *reduce_axis) const {
-    size_t x_rank = x_shape.size();
-    size_t y_rank = y_shape.size();
-
-    // Get reduc aixs and
-    int64_t compare_offset = x_rank - y_rank;
-
-    for (size_t i = 0; i < y_rank; ++i) {
-      if (y_shape[i] == 1 && x_shape[i + compare_offset] != 1) {
-        reduce_axis->push_back(compare_offset + i);
-      }
-    }
-  }
-  bool ProcessDynamicShape(paddle::dialect::ReduceAsOp op,
-                           std::vector<int64_t> *reduce_axis,
-                           std::vector<symbol::DimExpr> *output_dims,
-                           bool *x_y_shape_equal) const {
-    auto &shape_analysis =
-        pir::ShapeAnalysisManager::Instance().Get(op->GetParentProgram());
-
-    const auto &x_shape =
-        shape_analysis.GetShapeOrDataForValue(op->operand_source(0)).shape();
-    const auto &y_shape =
-        shape_analysis.GetShapeOrDataForValue(op->operand_source(1)).shape();
-
-    if (x_shape == y_shape) {
-      *x_y_shape_equal = true;
-      return true;
-    } else {
-      size_t x_rank = x_shape.size();
-      size_t y_rank = y_shape.size();
-
-      int64_t compare_offset = x_rank - y_rank;
-      bool can_replace_with_sum = true;
-      for (int64_t i = 0; i < compare_offset; ++i) {
-        output_dims->push_back(symbol::DimExpr(1));
-      }
-
-      for (size_t i = 0; i < y_rank; ++i) {
-        bool x_dim_i_eq_one = x_shape[i + compare_offset].isa<int64_t>() &&
-                              x_shape[i + compare_offset].Get<int64_t>() == 1;
-        bool y_dim_i_eq_one =
-            y_shape[i].isa<int64_t>() && y_shape[i].Get<int64_t>() == 1;
-        if (y_dim_i_eq_one && (!x_dim_i_eq_one)) {
-          reduce_axis->push_back(compare_offset + i);
-          output_dims->push_back(symbol::DimExpr(1));
-        } else if (x_shape[i + compare_offset] != y_shape[i]) {
-          can_replace_with_sum = false;
-          break;
-        } else {
-          output_dims->push_back(y_shape[i]);
-        }
-      }
-      return can_replace_with_sum;
-    }
-  }
-};
-
 PdOpToCinnOpPass::PdOpToCinnOpPass()
     : pir::PatternRewritePass("pd_to_cinn_pass", 1) {}
 
@@ -1400,7 +1244,6 @@ pir::RewritePatternSet PdOpToCinnOpPass::InitializePatterns(
   ps.Add<SigmoidOpPattern>(context);
   ps.Add<GatherOpPattern>(context);
   ps.Add<FlattenOpPattern>(context);
-  ps.Add<ReduceAsOpPattern>(context);
 
   return ps;
 }
